feat(config): add folder command
Adds a new `config:folder` command that checks for the existence of the `PROCA_CONFIG_FOLDER` and creates a set of predefined subdirectories if they are not present.

This command is useful for ensuring that the required directory structure is in place for the application to function correctly.

The command will output a list of the folders checked, indicating whether they were created or already existed.